add user logout to ctrl, simple script
the script in _layout doesn't properly route to a viable address, and it doesn't show a timer or anything to that effect. And the addition in program.cs doesn't mesh well with either of those yet. however the controller action works very well :^)

Not tested code, not sure if the timer gets refreshed on a new page

and most damning of all, the logout page doesnt log you out, but asks you to press a button to log out, gotta fix that ASAP